create base class for processing context, move shared code between fetchini and fetchtables into base class, and add functionality for interactive prompting of password
