+++ esgprep/fetchini/context.py
@@ -9,17 +9,15 @@
 
 import logging
 import os
-import re
 import sys
 
-from requests.auth import HTTPBasicAuth
 from tqdm import tqdm
 
 from constants import *
-from esgprep.utils.misc import gh_request_content
+from esgprep.utils.ctx_base import BaseContext
 
 
-class ProcessingContext(object):
+class ProcessingContext(BaseContext):
     """
     Encapsulates the processing context/information for main process.
 
@@ -49,7 +47,10 @@ class ProcessingContext(object):
         # Init output directory
         self.make_ini_dir()
         # Init the project list to retrieve
-        self.targets = self.target_projects()
+        self.targets = self.target_projects(
+            pattern = 'esg\.(.+?)\.ini',
+            url_format = self.url.format('')
+            )
         # Init progress bar
         nfiles = len(self.targets)
         if self.pbar and nfiles:
@@ -67,16 +68,6 @@ class ProcessingContext(object):
         # Default is sys.exit(0)
         if self.error:
             sys.exit(1)
-
-    def authenticate(self):
-        """
-        Builds GitHub HTTP authenticator
-
-        :returns: The HTTP authenticator
-        :rtype: *requests.auth.HTTPBasicAuth*
-
-        """
-        return HTTPBasicAuth(self.gh_user, self.gh_password) if self.gh_user and self.gh_password else None
 
     def make_ini_dir(self):
         """
@@ -98,23 +89,3 @@ class ProcessingContext(object):
                 if not os.path.isdir(self.config_dir):
                     os.makedirs(self.config_dir)
                     logging.warning('{} created'.format(self.config_dir))
-
-    def target_projects(self):
-        """
-        Gets the available projects ids from GitHub esg.*.ini files.
-        Make the intersection with the desired projects to fetch.
-
-        :returns: The target projects
-        :rtype: *list*
-
-        """
-        pattern = 'esg\.(.+?)\.ini'
-        r = gh_request_content(self.url.format(''), auth=self.auth)
-        files = [content['name'] for content in r.json()]
-        p_avail = set([re.search(pattern, x).group(1) for x in files if re.search(pattern, x)])
-        if self.projects:
-            p = set(self.projects)
-            p_avail = p_avail.intersection(p)
-            if p.difference(p_avail):
-                logging.warning("Unavailable project(s): {}".format(', '.join(p.difference(p_avail))))
-        return list(p_avail)
